# app/api/predict_api.go
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"log"
	"math"
	"net/http"
	"strings"
	"time"

	_client "github.com/Publikey/runqy/client"
	"github.com/Publikey/runqy/models"
	queueworker "github.com/Publikey/runqy/queues"
	"github.com/Publikey/runqy/utilities"
	"github.com/gin-gonic/gin"
	"github.com/hibiken/asynq"
	"github.com/redis/go-redis/v9"
)

// GetPredictStatus godoc
//
//	@Summary		Get the information about the task in the queue and it's response if any.
//	@Description	Retrieve the body of the response or the status of the request if has not been processed already.
//	@Tags			queue
//	@Accept			json
//	@Produce		json
//	@Param			uuid		path		string	true	" the uuid of the task returned from the POST query"
//	@Param			priority	path		string	true	"`normal` or `priority`"
//	@Success		200			{object}	models.ResponseGet
//	@Failure		400			{object}	models.APIErrorResponse
//	@Router			/queue/{uuid}/{priority} [get]
func GetPredictStatus(c *gin.Context) {
	uuid := c.Param("uuid")
	priority := c.Param("priority")

	redisAddr, err := models.BuildRedisConns()
	if err != nil {
		log.Fatalf("[FATAL] Redis build connection failed: %v", err)
	}

	inspector := asynq.NewInspector(redisAddr.AsynqOpt)
	defer inspector.Close()

	if priority == "priority" {
		resp, err := waitForResult(context.Background(), inspector, "priority", uuid)
		if err != nil {
			c.JSON(http.StatusBadRequest, models.APIErrorResponse{Errors: []string{err.Error()}})
			return
		}
		// Manual mapping from *asynq.TaskInfo to TaskInfoDoc
		taskDoc := models.GetTaskInfoDoc{
			ID:            resp.ID,
			Type:          resp.Type,
			Payload:       utilities.DecodeBase64OrReturnRaw(resp.Payload),
			State:         resp.State.String(),
			Queue:         resp.Queue,
			MaxRetry:      resp.MaxRetry,
			Retried:       resp.Retried,
			LastErr:       resp.LastErr,
			LastFailedAt:  resp.LastFailedAt,
			Deadline:      resp.Deadline,
			Group:         resp.Group,
			NextProcessAt: resp.NextProcessAt,
			IsOrphaned:    resp.IsOrphaned,
			CompletedAt:   resp.CompletedAt,
			Result:        utilities.DecodeBase64OrReturnRaw(resp.Result),
		}
		response := models.ResponseGet{
			Info: taskDoc,
		}
		c.JSON(http.StatusOK, response)
	} else {
		resp, err := waitForResult(context.Background(), inspector, "normal", uuid)
		if err != nil {
			c.JSON(http.StatusBadRequest, models.APIErrorResponse{Errors: []string{err.Error()}})
			return
		}
		// Manual mapping from *asynq.TaskInfo to TaskInfoDoc
		taskDoc := models.GetTaskInfoDoc{
			ID:            resp.ID,
			Type:          resp.Type,
			Payload:       utilities.DecodeBase64OrReturnRaw(resp.Payload),
			State:         resp.State.String(),
			Queue:         resp.Queue,
			MaxRetry:      resp.MaxRetry,
			Retried:       resp.Retried,
			LastErr:       resp.LastErr,
			LastFailedAt:  resp.LastFailedAt,
			Deadline:      resp.Deadline,
			Group:         resp.Group,
			NextProcessAt: resp.NextProcessAt,
			IsOrphaned:    resp.IsOrphaned,
			CompletedAt:   resp.CompletedAt,
			Result:        utilities.DecodeBase64OrReturnRaw(resp.Result),
		}
		response := models.ResponseGet{
			Info: taskDoc,
		}
		c.JSON(http.StatusOK, response)
	}

}

// NewPredict godoc
//
//	@Summary		Send a new task to the queuer
//	@Description	Send a generic task request with queue, timeout, and a flexible JSON payload.
//	@Tags			queue
//	@Accept			json
//	@Produce		json
//	@Param			task	body		models.GenericTask	true	"Task with queue, timeout, and payload"
//	@Success		200		{object}	models.GenericResponsePost
//	@Failure		400		{object}	models.APIErrorResponse
//	@Router			/queue/add [post]
//
// NewPredict returns a handler that validates the incoming task `data` against
// the queue worker YAML schemas found in `qwConfigDir` before enqueuing.
func NewPredict(qwConfigDir string) gin.HandlerFunc {
	return func(c *gin.Context) {
		query := models.GenericTask{}
		if err := c.ShouldBindBodyWithJSON(&query); err != nil {
			c.JSON(http.StatusBadRequest, models.APIErrorResponse{Errors: []string{err.Error()}})
			return
		}

		if query.Queue == "" {
			c.JSON(http.StatusBadRequest, models.APIErrorResponse{Errors: []string{"queue is required"}})
			return
		}

		// Prepare payloadToSend (TypedPayload by default)
		var payloadToSend interface{}

		// Validate payload against YAML schema for this queue if available
		yamls, err := queueworker.LoadAll(qwConfigDir)
		if err == nil && len(yamls) > 0 {
			var matched *queueworker.QueueYAML
			// Try to find the queue by matching runtime config names
			for _, y := range yamls {
				for baseName, qcfg := range y.Queues {
					configs := qcfg.ToQueueConfigs(baseName)
					for _, cfg := range configs {
						if cfg.Name == query.Queue {
							matched = &qcfg
							break
						}
					}
					if matched != nil {
						break
					}
				}
				if matched != nil {
					break
				}
			}

			// If not found, try trimming a trailing _default
			if matched == nil && strings.HasSuffix(query.Queue, "_default") {
				base := strings.TrimSuffix(query.Queue, "_default")
				for _, y := range yamls {
					if qcfg, ok := y.Queues[base]; ok {
						matched = &qcfg
						break
					}
				}
			}

			if matched != nil && len(matched.Input) > 0 {
				var dataMap map[string]interface{}
				if err := json.Unmarshal(query.Data, &dataMap); err != nil {
					c.JSON(http.StatusBadRequest, models.APIErrorResponse{Errors: []string{"invalid data payload: " + err.Error()}})
					return
				}

				// Validate required fields and types, and build filtered payload
				filtered := make(map[string]interface{})
				for _, field := range matched.Input {
					val, ok := dataMap[field.Name]
					if !ok {
						c.JSON(http.StatusBadRequest, models.APIErrorResponse{Errors: []string{fmt.Sprintf("%s is required", field.Name)}})
						return
					}

					if !checkAllowedType(val, field.Type) {
						c.JSON(http.StatusBadRequest, models.APIErrorResponse{Errors: []string{fmt.Sprintf("%s has invalid type", field.Name)}})
						return
					}

					// If the field expects an int but JSON gave float64, convert to int64
					if contains(field.Type, "int") {
						if f, ok := val.(float64); ok {
							// use int64 for integers
							filtered[field.Name] = int64(f)
							continue
						}
					}
					filtered[field.Name] = val
				}

				// Build typed payload (flat map) to send
				payloadToSend = models.TypedPayload(filtered)
			}
			// if matched == nil, fallthrough and parse into TypedPayload below
		}

		// If payloadToSend is still nil, create a TypedPayload from the raw data
		if payloadToSend == nil {
			var dataMap map[string]interface{}
			if err := json.Unmarshal(query.Data, &dataMap); err != nil {
				// If unmarshalling fails, fallback to sending raw bytes
				payloadToSend = json.RawMessage(query.Data)
			} else {
				payloadToSend = models.TypedPayload(dataMap)
			}
		}

		asynqClient := c.Keys["client"].(*asynq.Client)
		rdb := c.Keys["rdb"].(*redis.Client)

		info, err := _client.EnqueueGenericTask(asynqClient, rdb, query.Queue, query.Timeout, payloadToSend)
		if err != nil {
			c.JSON(http.StatusBadRequest, models.APIErrorResponse{Errors: []string{err.Error()}})
			return
		}
		// Manual mapping from *asynq.TaskInfo to TaskInfoDoc
		taskDoc := models.AddTaskInfoDoc{
			ID:            info.ID,
			Type:          info.Type,
			Payload:       info.Payload,
			State:         info.State.String(),
			Queue:         info.Queue,
			MaxRetry:      info.MaxRetry,
			Retried:       info.Retried,
			LastErr:       info.LastErr,
			LastFailedAt:  info.LastFailedAt,
			Deadline:      info.Deadline,
			Group:         info.Group,
			NextProcessAt: info.NextProcessAt,
			IsOrphaned:    info.IsOrphaned,
			CompletedAt:   info.CompletedAt,
			Result:        info.Result,
		}
		response := models.GenericResponsePost{
			Info: taskDoc,
			Data: query.Data,
		}
		c.JSON(http.StatusOK, response)
	}

}

func waitForResult(ctx context.Context, i *asynq.Inspector, queue, taskID string) (*asynq.TaskInfo, error) {
	t := time.NewTicker(time.Second)
	defer t.Stop()
	for {
		select {
		case <-t.C:
			taskInfo, err := i.GetTaskInfo(queue, taskID)
			if err != nil {
				return nil, err
			}
			if taskInfo.LastErr != "" {
				return taskInfo, nil
			}
			return taskInfo, nil
		case <-ctx.Done():
			return nil, fmt.Errorf("context closed")
		}
	}
}

// checkAllowedType checks if the provided value matches any of the allowed types
func checkAllowedType(v interface{}, allowed []string) bool {
	switch val := v.(type) {
	case string:
		return contains(allowed, "string")
	case bool:
		return contains(allowed, "bool")
	case float64:
		// JSON numbers are float64; determine if integer-valued
		if math.Trunc(val) == val {
			return contains(allowed, "int") || contains(allowed, "float")
		}
		return contains(allowed, "float")
	case []interface{}:
		return contains(allowed, "array")
	case map[string]interface{}:
		return contains(allowed, "object")
	default:
		return false
	}
}

func contains(arr []string, s string) bool {
	for _, v := range arr {
		if v == s {
			return true
		}
	}
	return false
}