# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Task queueing service built on asynq for managing distributed task processing",
        "title": "Runqy Queue API",
        "version": "1.0.0"
    },
    "paths": {
        "/queue/add": {
            "post": {
                "description": "Send a prediction request. Set `provider` to route to different backends:\n- `\"\"` or `\"worker\"`: External workers (SDXL) - uses num_steps, guidance_scale, seed, negative_prompt, scheduler, model_name\n- `\"azure\"`: Azure OpenAI DALL-E - uses quality (\"hd\"/\"standard\")\nCommon fields: prompt (required), width, height, batch_nbr. Unused fields are ignored.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "Send a new predict to the queuer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The type of the generation either image or video",
                        "name": "type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "{object}",
                        "name": "predict",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Predict"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponsePost"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/queue/{uuid}/{priority}": {
            "get": {
                "description": "Retrieve the body of the response or the status of the request if has not been processed already.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "Get the information about the task in the queue and it's response if any.",
                "parameters": [
                    {
                        "type": "string",
                        "description": " the uuid of the task returned from the POST query",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "`normal` or `priority`",
                        "name": "priority",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseGet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.AddTaskInfoDoc": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isOrphaned": {
                    "type": "boolean"
                },
                "lastErr": {
                    "type": "string"
                },
                "lastFailedAt": {
                    "type": "string"
                },
                "max_retry": {
                    "type": "integer"
                },
                "nextProcessAt": {
                    "type": "string"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "queue": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "retried": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.GetTaskInfoDoc": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isOrphaned": {
                    "type": "boolean"
                },
                "lastErr": {
                    "type": "string"
                },
                "lastFailedAt": {
                    "type": "string"
                },
                "max_retry": {
                    "type": "integer"
                },
                "nextProcessAt": {
                    "type": "string"
                },
                "payload": {
                    "type": "string"
                },
                "queue": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "retried": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Predict": {
            "type": "object",
            "properties": {
                "enqueued_at": {
                    "type": "integer"
                },
                "model_input": {
                    "$ref": "#/definitions/models.Request"
                },
                "predict_id": {
                    "type": "integer"
                },
                "provider": {
                    "description": "\"azure\", \"openai\", \"google\", \"worker\" or empty (default: worker)",
                    "type": "string"
                },
                "queue": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timeout": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "webhook_url": {
                    "type": "string"
                }
            }
        },
        "models.Request": {
            "type": "object",
            "properties": {
                "batch_nbr": {
                    "description": "BatchNbr is the number of images to generate (all providers)",
                    "type": "integer"
                },
                "clip_skip": {
                    "description": "ClipSkip is the number of CLIP layers to skip (worker/SDXL only)",
                    "type": "integer"
                },
                "decode_noise_scale": {
                    "description": "DecodeNoiseScale for video decode process (worker/SDXL only)",
                    "type": "number"
                },
                "decode_timesteps": {
                    "description": "DecodeTimesteps for video decode process (worker/SDXL only)",
                    "type": "number"
                },
                "fps": {
                    "description": "FPS is the frames per second for video output (worker/SDXL only)",
                    "type": "integer"
                },
                "guidance_scale": {
                    "description": "GuidanceScale controls how closely to follow the prompt (worker/SDXL only)",
                    "type": "number"
                },
                "height": {
                    "description": "Height of the output image in pixels (all providers)",
                    "type": "integer"
                },
                "id": {
                    "description": "ID is the request identifier (all providers)",
                    "type": "string"
                },
                "img_url": {
                    "description": "ImgURL is the input image URL for img2img generation (worker/SDXL only)",
                    "type": "string"
                },
                "model_name": {
                    "description": "ModelName is the model to use. Worker: SDXL model name. Azure: \"gpt-image-1\", \"gpt-image-1.5\", \"flux.2-pro\", \"flux-1.1-pro\" (required)",
                    "type": "string"
                },
                "negative_prompt": {
                    "description": "NegativePrompt specifies what to avoid in generation (worker/SDXL only)",
                    "type": "string"
                },
                "num_frames": {
                    "description": "NumFrames is the number of frames for video generation (worker/SDXL only)",
                    "type": "integer"
                },
                "num_steps": {
                    "description": "NumSteps is the number of diffusion steps (worker/SDXL only)",
                    "type": "integer"
                },
                "output_compression": {
                    "description": "OutputCompression controls image compression 0-100 (Azure: gpt-image-1, gpt-image-1.5 only)",
                    "type": "integer"
                },
                "output_format": {
                    "description": "OutputFormat specifies output format: \"png\", \"jpeg\" (Azure: gpt-image-1, gpt-image-1.5, flux-1.1-pro)",
                    "type": "string"
                },
                "prompt": {
                    "description": "Prompt is the text prompt for generation - REQUIRED (all providers)",
                    "type": "string"
                },
                "quality": {
                    "description": "Quality level: \"hd\", \"standard\" (Azure: dall-e-3, gpt-image-1 only - not flux)",
                    "type": "string"
                },
                "scheduler": {
                    "description": "Scheduler is the sampling scheduler to use (worker/SDXL only)",
                    "type": "string"
                },
                "seed": {
                    "description": "Seed for reproducible generation (worker/SDXL only)",
                    "type": "integer"
                },
                "webhook_url": {
                    "description": "URL is an alternative webhook URL field (all providers)",
                    "type": "string"
                },
                "width": {
                    "description": "Width of the output image in pixels (all providers)",
                    "type": "integer"
                }
            }
        },
        "models.ResponseGet": {
            "type": "object",
            "properties": {
                "info": {
                    "$ref": "#/definitions/models.GetTaskInfoDoc"
                }
            }
        },
        "models.ResponsePost": {
            "type": "object",
            "properties": {
                "info": {
                    "$ref": "#/definitions/models.AddTaskInfoDoc"
                },
                "userPredict": {
                    "$ref": "#/definitions/models.Predict"
                }
            }
        }
    }
}